Implement authorization using Gates

+++ app/Http/Controllers/QuestionsController.php
@@ -65,7 +65,12 @@ class QuestionsController extends Controller
      */
     public function edit(Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access denied");
+        }
+
         return view('questions.edit', compact('question'));
+        
     }
 
     /**
@@ -77,6 +82,10 @@ class QuestionsController extends Controller
      */
     public function update(AskQuestionRequest $request, Question $question)
     {
+        if(\Gate::denies('update-question', $question)){
+            abort(403, "Access denied");
+        }
+
         $question->update($request->only('title', 'body'));
 
         return redirect()->route('questions.index')->with('success', 'Question has been updated.');
@@ -90,8 +99,11 @@ class QuestionsController extends Controller
      */
     public function destroy(Question $question)
     {
-        $question->delete();
+        if(\Gate::denies('delete-question', $question)){
+            abort(403, "Access denied");
+        }
 
+        $question->delete();
         return redirect()->route('questions.index')->with('success', "Question has been deleted.");
 
     }
